feat(user): 도메인 별로 예외 클래스를 분리한다.
